Support: seperate internal dependencies definitions
- Seperate the interface definitions of dependencies into a
  separate file, to clean up controller.go.

- Update the Comm interface with two Send* methods
