+++ internal/bft/controller.go
@@ -16,24 +16,6 @@ import (
 	protos "github.com/SmartBFT-Go/consensus/smartbftprotos"
 )
 
-//go:generate mockery -dir . -name Verifier -case underscore -output ./mocks/
-type Verifier interface {
-	VerifyProposal(proposal types.Proposal, prevHeader []byte) ([]types.RequestInfo, error)
-	VerifyRequest(val []byte) (types.RequestInfo, error)
-	VerifyConsenterSig(signature types.Signature, prop types.Proposal) error
-	VerificationSequence() uint64
-}
-
-//go:generate mockery -dir . -name Assembler -case underscore -output ./mocks/
-type Assembler interface {
-	AssembleProposal(metadata []byte, requests [][]byte) (nextProp types.Proposal, remainder [][]byte)
-}
-
-//go:generate mockery -dir . -name Application -case underscore -output ./mocks/
-type Application interface {
-	Deliver(proposal types.Proposal, signature []types.Signature)
-}
-
 //go:generate mockery -dir . -name Decider -case underscore -output ./mocks/
 type Decider interface {
 	Decide(proposal types.Proposal, signatures []types.Signature, requests []types.RequestInfo)
@@ -42,22 +24,6 @@ type Decider interface {
 //go:generate mockery -dir . -name FailureDetector -case underscore -output ./mocks/
 type FailureDetector interface {
 	Complain()
-}
-
-//go:generate mockery -dir . -name Synchronizer -case underscore -output ./mocks/
-type Synchronizer interface {
-	Sync() (protos.ViewMetadata, uint64)
-}
-
-//go:generate mockery -dir . -name Comm -case underscore -output ./mocks/
-type Comm interface {
-	BroadcastConsensus(m *protos.Message)
-}
-
-//go:generate mockery -dir . -name Signer -case underscore -output ./mocks/
-type Signer interface {
-	Sign([]byte) []byte
-	SignProposal(types.Proposal) *types.Signature
 }
 
 //go:generate mockery -dir . -name Batcher -case underscore -output ./mocks/
